+++ client/src/components/Account/EditEvent.js
@@ -1,7 +1,7 @@
 import React, { useState, useEffect } from "react";
 import { ToastContainer, toast } from 'react-toastify';
 import 'react-toastify/dist/ReactToastify.css';
-import { updateEvent } from "../../utils/API";
+import { updateEvent } from "../utils/API";
 
 
 // Toast Notification Function to let the user know the event was deleted
@@ -32,10 +32,10 @@ const EditEvent = (props) => {
     setNewEditEvent(editEventForm);
   }, []);
 
-const formStyles = {
-  margin: "10px"
-}
-  
+  const formStyles = {
+    margin: "10px"
+  }
+
   // Return the component
   return (
     // Div to center the content
@@ -96,7 +96,7 @@ const formStyles = {
 
           {/* Button to save the event and call the Toast Notify Function*/}
           <button
-            onClick={() => { updateEvent(newEditEvent); notify(newEditEvent.name)}}
+            onClick={() => { updateEvent(newEditEvent); notify(newEditEvent.name) }}
             className="justify-content-md-center col-md-12 btn btn-primary"
             style={formStyles}
           >
