+++ client/src/components/Map/Details.js
@@ -1,10 +1,10 @@
 import React, { useState } from "react";
-import { deleteEvent } from "../../utils/API";
+import { deleteEvent } from "../utils/API";
 import DefaultDetails from "./DefaultDetails";
 import "./Details.css";
 import { ToastContainer, toast } from "react-toastify";
 import "react-toastify/dist/ReactToastify.css";
-import { updateEvent } from "../../utils/API";
+import { updateEvent } from "../utils/API";
 
 // Toast Notification Function to let the user know the event was deleted
 const notify = () => toast("Event Deleted");
